+++ components/AnimationContainer.tsx
