+++ components/Guide.tsx
