+++ pages/join/nickname.tsx
@@ -7,9 +7,9 @@ import { checkIdValid } from '../../util/checkIdValid';
 
 import Top from '../../components/Top';
 import ProgressBar from '../../components/ProgressBar';
-import Container from '../../components/Join/Container';
-import InputContainer from '../../components/Join/InputContainer';
-import Guide from '../../components/Join/Guide';
+import Container from '../../components/Container';
+import InputContainer from '../../components/AnimationContainer';
+import Guide from '../../components/Guide';
 import { StyledInput } from '../../components/Join/FormElement';
 import InputMessage from '../../components/Join/InputMessage';
 import Button from '../../components/Button';
