+++ pages/join/school.tsx
@@ -7,9 +7,9 @@ import { userInfoState } from '../../core/recoil/userInfoAtom';
 
 import Top from '../../components/Top';
 import ProgressBar from '../../components/ProgressBar';
-import Container from '../../components/Join/Container';
-import InputContainer from '../../components/Join/InputContainer';
-import Guide from '../../components/Join/Guide';
+import Container from '../../components/Container';
+import InputContainer from '../../components/AnimationContainer';
+import Guide from '../../components/Guide';
 import { StyledInput } from '../../components/Join/FormElement';
 import Button from '../../components/Button';
 import { ValidateButton } from '../../components/Join/FormElement';
